_requires_grad -> requires_grad
This makes it match what eager mode uses as the keyword name.
Currently _requires_grad will not appear in serialization because
it is not listed as kwarg only. There is a small chance there is a model
that has never been run in eager mode that uses the _requires_grad name,
but this is rare enough that I don't think we need to worry about it unless
something breaks in testing.

ghstack-source-id: 500768645914aa4216c51bd9ddaf9555c623a8d1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36168